Create common AWS config
