+++ metadata-ingestion/src/datahub/ingestion/source/glue.py
@@ -4,17 +4,13 @@ import typing
 from collections import defaultdict
 from dataclasses import dataclass
 from dataclasses import field as dataclass_field
-from functools import reduce
 from typing import Any, Dict, Iterable, Iterator, List, Optional, Set, Tuple, Union
 from urllib.parse import urlparse
 
-import boto3
-
-from datahub.configuration import ConfigModel
-from datahub.configuration.common import AllowDenyPattern
 from datahub.emitter import mce_builder
 from datahub.ingestion.api.common import PipelineContext
 from datahub.ingestion.api.source import Source, SourceReport
+from datahub.ingestion.source.aws_common import AwsSourceConfig
 from datahub.ingestion.source.metadata_common import MetadataWorkUnit
 from datahub.metadata.com.linkedin.pegasus2avro.common import AuditStamp, Status
 from datahub.metadata.com.linkedin.pegasus2avro.metadata.snapshot import DatasetSnapshot
@@ -50,86 +46,9 @@ from datahub.metadata.schema_classes import (
 )
 
 
-def assume_role(
-    role_arn: str, aws_region: str, credentials: Optional[dict] = None
-) -> dict:
-    credentials = credentials or {}
-    sts_client = boto3.client(
-        "sts",
-        region_name=aws_region,
-        aws_access_key_id=credentials.get("AccessKeyId"),
-        aws_secret_access_key=credentials.get("SecretAccessKey"),
-        aws_session_token=credentials.get("SessionToken"),
-    )
-
-    assumed_role_object = sts_client.assume_role(
-        RoleArn=role_arn, RoleSessionName="DatahubIngestionSourceGlue"
-    )
-    return assumed_role_object["Credentials"]
-
-
-class GlueSourceConfig(ConfigModel):
-    env: str = "PROD"
-
-    database_pattern: AllowDenyPattern = AllowDenyPattern.allow_all()
-    table_pattern: AllowDenyPattern = AllowDenyPattern.allow_all()
+class GlueSourceConfig(AwsSourceConfig):
 
     extract_transforms: Optional[bool] = True
-
-    aws_access_key_id: Optional[str] = None
-    aws_secret_access_key: Optional[str] = None
-    aws_session_token: Optional[str] = None
-    aws_role: Optional[Union[str, List[str]]] = None
-    aws_region: str
-
-    def get_client(self, service: str) -> boto3.client:
-        if (
-            self.aws_access_key_id
-            and self.aws_secret_access_key
-            and self.aws_session_token
-        ):
-            return boto3.client(
-                service,
-                aws_access_key_id=self.aws_access_key_id,
-                aws_secret_access_key=self.aws_secret_access_key,
-                aws_session_token=self.aws_session_token,
-                region_name=self.aws_region,
-            )
-        elif self.aws_access_key_id and self.aws_secret_access_key:
-            return boto3.client(
-                service,
-                aws_access_key_id=self.aws_access_key_id,
-                aws_secret_access_key=self.aws_secret_access_key,
-                region_name=self.aws_region,
-            )
-        elif self.aws_role:
-            if isinstance(self.aws_role, str):
-                credentials = assume_role(self.aws_role, self.aws_region)
-            else:
-                credentials = reduce(
-                    lambda new_credentials, role_arn: assume_role(
-                        role_arn, self.aws_region, new_credentials
-                    ),
-                    self.aws_role,
-                    {},
-                )
-            return boto3.client(
-                service,
-                aws_access_key_id=credentials["AccessKeyId"],
-                aws_secret_access_key=credentials["SecretAccessKey"],
-                aws_session_token=credentials["SessionToken"],
-                region_name=self.aws_region,
-            )
-        else:
-            return boto3.client(service, region_name=self.aws_region)
-
-    @property
-    def glue_client(self):
-        return self.get_client("glue")
-
-    @property
-    def s3_client(self):
-        return self.get_client("s3")
 
 
 @dataclass
